remove duplicate tests
