+++ test/integration/api/allocation/test.bats
@@ -18,6 +18,10 @@ teardown() {
    go test valid_total_efficiency_test.go
 }
 
+@test "validate_api: validate Total cost vs individual cost" {
+   go test individual_costs_sum_vs_totalcost_test.go
+}
+
 @test "validate_api: validate Window Start and End" {
     go test correct_window_values_test.go
 }
